# SPU: Use REP MOVSB in do_dma_transfer