Modern implementations of memcpy will use the REP MOVSB instruction when the size of the copy is large, and when the CPU is known to have a fast implementation of the instruction. The speed of the instruction is checked through 2 bits on cpuid, ERMS (enhanced rep movsb) and FSRM (fast short rep movsb). Support for detecting these flags has been added, and you will see in your log whether your CPU supports these flags or not.

On intel ERMS is supported as far back as ivy bridge, while FSRM is supported on icelake and up.
On amd, only the latest zen3 cpus have either (zen3 has both ERMS and FSRM)

Because I don't have access to every cpu out there, I based the thresholds for when rep movsb is preferred over simd copies roughly off of glibc.

On my 7700K, I couldn't measure any positive or negative performance impact.
On my 1135G7, there was a small performance uplift, and measurably less time spent in do_dma_transfer according to profiling.

<s>Need testing to determine:
- That it doesn't nuke performance on any cpus which don't support either ERMS or FSRM
- That it doesn't nuke performance on any cpus which do support either ERMS or FSRM
- That it improves performance on anything other than my tigerlake laptop
- That it doesn't break anything</s>

If it doesn't seem to really help anything outside of my laptop, then we can just close the PR. If I didn't have access to this laptop then I probably wouldn't have PR'd this in the first place.