@Yahfz 2021-10-30
If there was any difference in performance its too little to be measured. I've done several runs in god of war 3 and sometimes one would beat the other. All the tests were within 1-2fps of each other on average.