@elad335 2021-11-01
Has anyone tried memcpy recently? It's been a couple of years since it was last tested could have been optimized.