@RipleyTom 2021-11-08
I added a 16k PUT test in spurs_test and tested this vs master:
On 3950X:

PR(rebased on master):
```
Large Put Perf completed in 20165 ms (PS3: 0 ms)
Large Put Perf completed in 20397 ms (PS3: 0 ms)
Large Put Perf completed in 21206 ms (PS3: 0 ms)
Large Put Perf completed in 20975 ms (PS3: 0 ms)
Large Put Perf completed in 20802 ms (PS3: 0 ms)
```

Master:
```
Large Put Perf completed in 13231 ms (PS3: 0 ms)
Large Put Perf completed in 13128 ms (PS3: 0 ms)
Large Put Perf completed in 13000 ms (PS3: 0 ms)
Large Put Perf completed in 12855 ms (PS3: 0 ms)
Large Put Perf completed in 13447 ms (PS3: 0 ms)
```