@Jhonny-2005 2021-11-15
Many interesting observations about the implementation are described [here](https://stackoverflow.com/questions/43343231/enhanced-rep-movsb-for-memcpy)