@AniLeo 2021-11-20
R7 5800X

Master:
Large PUT Perf completed in 3210 ms (PS3: 4454 ms)
Large PUT Perf completed in 3256 ms (PS3: 4454 ms)
Large PUT Perf completed in 3260 ms (PS3: 4454 ms)

PR:
Large PUT Perf completed in 3094 ms (PS3: 4454 ms)
Large PUT Perf completed in 3103 ms (PS3: 4454 ms)
Large PUT Perf completed in 3103 ms (PS3: 4454 ms)

No measurable difference in God of War: Ascension and Mandelbrot