@YuriNator557 2021-12-31
3700x

PR
Large PUT Perf completed in 6554 ms (PS3: 4454 ms)
Large PUT Perf completed in 6435 ms (PS3: 4454 ms)
Large PUT Perf completed in 6435 ms (PS3: 4454 ms)

Master
Large PUT Perf completed in 7684 ms (PS3: 4454 ms)
Large PUT Perf completed in 7919 ms (PS3: 4454 ms)
Large PUT Perf completed in 7692 ms (PS3: 4454 ms)